cleaned up function names and file names as well as created transaction class dart file
