+++ crypto_balance/.flutter-plugins-dependencies
@@ -1,1 +1,1 @@
-{"info":"This is a generated file; do not edit or check into version control.","plugins":{"ios":[{"name":"permission_handler","path":"C:\\\\Users\\\\Scott\\\\AppData\\\\Local\\\\Pub\\\\Cache\\\\hosted\\\\pub.dartlang.org\\\\permission_handler-5.0.1+1\\\\","dependencies":[]},{"name":"sqflite","path":"C:\\\\Users\\\\Scott\\\\AppData\\\\Local\\\\Pub\\\\Cache\\\\hosted\\\\pub.dartlang.org\\\\sqflite-1.3.1+1\\\\","dependencies":[]},{"name":"url_launcher","path":"C:\\\\Users\\\\Scott\\\\AppData\\\\Local\\\\Pub\\\\Cache\\\\hosted\\\\pub.dartlang.org\\\\url_launcher-5.7.2\\\\","dependencies":[]}],"android":[{"name":"flutter_plugin_android_lifecycle","path":"C:\\\\Users\\\\Scott\\\\AppData\\\\Local\\\\Pub\\\\Cache\\\\hosted\\\\pub.dartlang.org\\\\flutter_plugin_android_lifecycle-1.0.11\\\\","dependencies":[]},{"name":"permission_handler","path":"C:\\\\Users\\\\Scott\\\\AppData\\\\Local\\\\Pub\\\\Cache\\\\hosted\\\\pub.dartlang.org\\\\permission_handler-5.0.1+1\\\\","dependencies":[]},{"name":"sqflite","path":"C:\\\\Users\\\\Scott\\\\AppData\\\\Local\\\\Pub\\\\Cache\\\\hosted\\\\pub.dartlang.org\\\\sqflite-1.3.1+1\\\\","dependencies":[]},{"name":"url_launcher","path":"C:\\\\Users\\\\Scott\\\\AppData\\\\Local\\\\Pub\\\\Cache\\\\hosted\\\\pub.dartlang.org\\\\url_launcher-5.7.2\\\\","dependencies":[]}],"macos":[{"name":"sqflite","path":"C:\\\\Users\\\\Scott\\\\AppData\\\\Local\\\\Pub\\\\Cache\\\\hosted\\\\pub.dartlang.org\\\\sqflite-1.3.1+1\\\\","dependencies":[]},{"name":"url_launcher_macos","path":"C:\\\\Users\\\\Scott\\\\AppData\\\\Local\\\\Pub\\\\Cache\\\\hosted\\\\pub.dartlang.org\\\\url_launcher_macos-0.0.1+8\\\\","dependencies":[]}],"linux":[{"name":"url_launcher_linux","path":"C:\\\\Users\\\\Scott\\\\AppData\\\\Local\\\\Pub\\\\Cache\\\\hosted\\\\pub.dartlang.org\\\\url_launcher_linux-0.0.1+1\\\\","dependencies":[]}],"windows":[{"name":"url_launcher_windows","path":"C:\\\\Users\\\\Scott\\\\AppData\\\\Local\\\\Pub\\\\Cache\\\\hosted\\\\pub.dartlang.org\\\\url_launcher_windows-0.0.1+1\\\\","dependencies":[]}],"web":[{"name":"url_launcher_web","path":"C:\\\\Users\\\\Scott\\\\AppData\\\\Local\\\\Pub\\\\Cache\\\\hosted\\\\pub.dartlang.org\\\\url_launcher_web-0.1.4+1\\\\","dependencies":[]}]},"dependencyGraph":[{"name":"flutter_plugin_android_lifecycle","dependencies":[]},{"name":"permission_handler","dependencies":[]},{"name":"sqflite","dependencies":[]},{"name":"url_launcher","dependencies":["url_launcher_web","url_launcher_linux","url_launcher_macos","url_launcher_windows"]},{"name":"url_launcher_linux","dependencies":[]},{"name":"url_launcher_macos","dependencies":[]},{"name":"url_launcher_web","dependencies":[]},{"name":"url_launcher_windows","dependencies":[]}],"date_created":"2021-01-12 10:45:52.804744","version":"1.23.0-8.0.pre.42"}
+{"info":"This is a generated file; do not edit or check into version control.","plugins":{"ios":[{"name":"permission_handler","path":"C:\\\\Users\\\\Scott\\\\AppData\\\\Local\\\\Pub\\\\Cache\\\\hosted\\\\pub.dartlang.org\\\\permission_handler-5.0.1+1\\\\","dependencies":[]},{"name":"sqflite","path":"C:\\\\Users\\\\Scott\\\\AppData\\\\Local\\\\Pub\\\\Cache\\\\hosted\\\\pub.dartlang.org\\\\sqflite-1.3.1+1\\\\","dependencies":[]},{"name":"url_launcher","path":"C:\\\\Users\\\\Scott\\\\AppData\\\\Local\\\\Pub\\\\Cache\\\\hosted\\\\pub.dartlang.org\\\\url_launcher-5.7.2\\\\","dependencies":[]}],"android":[{"name":"flutter_plugin_android_lifecycle","path":"C:\\\\Users\\\\Scott\\\\AppData\\\\Local\\\\Pub\\\\Cache\\\\hosted\\\\pub.dartlang.org\\\\flutter_plugin_android_lifecycle-1.0.11\\\\","dependencies":[]},{"name":"permission_handler","path":"C:\\\\Users\\\\Scott\\\\AppData\\\\Local\\\\Pub\\\\Cache\\\\hosted\\\\pub.dartlang.org\\\\permission_handler-5.0.1+1\\\\","dependencies":[]},{"name":"sqflite","path":"C:\\\\Users\\\\Scott\\\\AppData\\\\Local\\\\Pub\\\\Cache\\\\hosted\\\\pub.dartlang.org\\\\sqflite-1.3.1+1\\\\","dependencies":[]},{"name":"url_launcher","path":"C:\\\\Users\\\\Scott\\\\AppData\\\\Local\\\\Pub\\\\Cache\\\\hosted\\\\pub.dartlang.org\\\\url_launcher-5.7.2\\\\","dependencies":[]}],"macos":[{"name":"sqflite","path":"C:\\\\Users\\\\Scott\\\\AppData\\\\Local\\\\Pub\\\\Cache\\\\hosted\\\\pub.dartlang.org\\\\sqflite-1.3.1+1\\\\","dependencies":[]},{"name":"url_launcher_macos","path":"C:\\\\Users\\\\Scott\\\\AppData\\\\Local\\\\Pub\\\\Cache\\\\hosted\\\\pub.dartlang.org\\\\url_launcher_macos-0.0.1+8\\\\","dependencies":[]}],"linux":[{"name":"url_launcher_linux","path":"C:\\\\Users\\\\Scott\\\\AppData\\\\Local\\\\Pub\\\\Cache\\\\hosted\\\\pub.dartlang.org\\\\url_launcher_linux-0.0.1+1\\\\","dependencies":[]}],"windows":[{"name":"url_launcher_windows","path":"C:\\\\Users\\\\Scott\\\\AppData\\\\Local\\\\Pub\\\\Cache\\\\hosted\\\\pub.dartlang.org\\\\url_launcher_windows-0.0.1+1\\\\","dependencies":[]}],"web":[{"name":"url_launcher_web","path":"C:\\\\Users\\\\Scott\\\\AppData\\\\Local\\\\Pub\\\\Cache\\\\hosted\\\\pub.dartlang.org\\\\url_launcher_web-0.1.4+1\\\\","dependencies":[]}]},"dependencyGraph":[{"name":"flutter_plugin_android_lifecycle","dependencies":[]},{"name":"permission_handler","dependencies":[]},{"name":"sqflite","dependencies":[]},{"name":"url_launcher","dependencies":["url_launcher_web","url_launcher_linux","url_launcher_macos","url_launcher_windows"]},{"name":"url_launcher_linux","dependencies":[]},{"name":"url_launcher_macos","dependencies":[]},{"name":"url_launcher_web","dependencies":[]},{"name":"url_launcher_windows","dependencies":[]}],"date_created":"2021-01-12 12:04:09.529635","version":"1.23.0-8.0.pre.42"}
